+++ concur.js
@@ -1,5 +1,5 @@
 /**
- * Concur 0.2.4 - https://github.com/insin/concur
+ * Concur 0.2.5 - https://github.com/insin/concur
  * MIT Licensed
  */
 !function(e){if("object"==typeof exports)module.exports=e();else if("function"==typeof define&&define.amd)define(e);else{var f;"undefined"!=typeof window?f=window:"undefined"!=typeof global?f=global:"undefined"!=typeof self&&(f=self),f.Concur=e()}}(function(){var define,module,exports;return (function e(t,n,r){function s(o,u){if(!n[o]){if(!t[o]){var a=typeof require=="function"&&require;if(!u&&a)return a(o,!0);if(i)return i(o,!0);throw new Error("Cannot find module '"+o+"'")}var f=n[o]={exports:{}};t[o][0].call(f.exports,function(e){var n=t[o][1][e];return s(n?n:e)},f,f.exports,e,t,n,r)}return n[o].exports}var i=typeof require=="function"&&require;for(var o=0;o<r.length;o++)s(r[o]);return s})({1:[function(require,module,exports){
@@ -45,17 +45,16 @@ function applyMixins(properties) {
  * If a child constructor is not provided via prototypeProps.constructor,
  * a new constructor will be created.
  */
-function inheritFrom(parentConstructor, prototypeProps, constructorProps) {
-  // Get or create a child constructor
-  var childConstructor
-  if (prototypeProps && object.hasOwn(prototypeProps, 'constructor')) {
-    childConstructor = prototypeProps.constructor
-  }
-  else {
+function inheritFrom(parentConstructor, childConstructor, prototypeProps, constructorProps) {
+  // Create a child constructor if one wasn't given
+  if (childConstructor == null) {
     childConstructor = function() {
       parentConstructor.apply(this, arguments)
     }
   }
+
+  // Make sure the new prototype has the correct constructor set up
+  prototypeProps.constructor = childConstructor
 
   // Base constructors should only have the properties they're defined with
   if (parentConstructor !== Concur) {
@@ -64,15 +63,14 @@ function inheritFrom(parentConstructor, prototypeProps, constructorProps) {
     childConstructor.__super__ = parentConstructor.prototype
   }
 
-  // Add prototype properties, if given
-  if (prototypeProps) {
-    object.extend(childConstructor.prototype, prototypeProps)
-  }
+  // Add prototype properties - this is why we took a copy of the child
+  // constructor reference in extend() - if a .constructor had been passed as a
+  // __mixin__ and overitten prototypeProps.constructor, these properties would
+  // be getting set on the mixed-in constructor's prototype.
+  object.extend(childConstructor.prototype, prototypeProps)
 
-  // Add constructor properties, if given
-  if (constructorProps) {
-    object.extend(childConstructor, constructorProps)
-  }
+  // Add constructor properties
+  object.extend(childConstructor, constructorProps)
 
   return childConstructor
 }
@@ -94,29 +92,34 @@ Concur.__mro__ = []
  * context, which is expected to be a constructor.
  */
 Concur.extend = function(prototypeProps, constructorProps) {
+  // Ensure we have prop objects to work with
+  prototypeProps = prototypeProps || {}
+  constructorProps = constructorProps || {}
+
   // If the constructor being inherited from has a __meta__ function somewhere
   // in its prototype chain, call it to customise prototype and constructor
   // properties before they're used to set up the new constructor's prototype.
   if (typeof this.prototype.__meta__ != 'undefined') {
-    // Property objects must always exist so properties can be added to
-    // and removed from them.
-    prototypeProps = prototypeProps || {}
-    constructorProps = constructorProps || {}
     this.prototype.__meta__(prototypeProps, constructorProps)
   }
 
+  // Any child constructor passed in should take precedence - grab a reference
+  // to it befoer we apply any mixins.
+  var childConstructor = object.get(prototypeProps, 'constructor', null)
+
   // If any mixins are specified, mix them into the property objects
-  if (prototypeProps && object.hasOwn(prototypeProps, '__mixin__')) {
+  if (object.hasOwn(prototypeProps, '__mixin__')) {
     prototypeProps = applyMixins(prototypeProps)
   }
-  if (constructorProps && object.hasOwn(constructorProps, '__mixin__')) {
+  if (object.hasOwn(constructorProps, '__mixin__')) {
     constructorProps = applyMixins(constructorProps)
   }
 
   // Set up the new child constructor and its prototype
-  var childConstructor = inheritFrom(this,
-                                     prototypeProps,
-                                     constructorProps)
+  childConstructor = inheritFrom(this,
+                                 childConstructor,
+                                 prototypeProps,
+                                 constructorProps)
 
   // Pass on the extend function for extension in turn
   childConstructor.extend = this.extend