+++ CHANGES.rst
@@ -1,3 +1,9 @@
+0.2.5 / 2014-03-07
+==================
+
+* Fixed ``.constructor`` function property on a ``__mixin__`` replacing the
+  intended constructor.
+
 0.2.4 / 2014-02-23
 ==================